requirements.txt
aiohttp==3.7.4.post0
aiohttp-retry==2.4.6
alembic==1.4.1
appdirs==1.4.4
arrow==1.2.0
async-timeout==3.0.1       
atomicwrites==1.4.0        
atpublic==2.3
attrs==21.2.0
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7  
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cookiecutter==1.7.3        
cryptography==35.0.0       
databricks-cli==0.16.2     
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
docker==5.0.3
dpath==2.0.5
dulwich==0.20.25
dvc==2.8.2
entrypoints==0.3
Flask==2.0.2
flatten-dict==0.4.2        
flufl.lock==6.0
fsspec==2021.10.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.8
GitPython==3.1.24
google-api-core==2.1.1     
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0googleapis-common-protos==1.53.0
grandalf==0.6
greenlet==1.1.2
httplib2==0.20.1
idna==3.3
importlib-metadata==4.8.1  
iniconfig==1.1.1
install==1.3.4
itsdangerous==2.0.1        
Jinja2==3.0.2
jinja2-time==0.2.0
joblib==1.1.0
mailchecker==4.0.13        
Mako==1.1.5
MarkupSafe==2.0.1
mlflow==1.20.2
multidict==5.2.0
nanotime==0.5.2
networkx==2.6.3
numpy==1.21.3
oauth2client==4.1.3        
packaging==21.0
pandas==1.3.4
pathspec==0.9.0
phonenumbers==8.12.35      
pluggy==1.0.0
ply==3.11
poyo==0.5.0
prometheus-client==0.11.0  
prometheus-flask-exporter==0.18.4
protobuf==3.19.0
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8      
pycparser==2.20
pydot==1.4.2
PyDrive2==1.10.0
pygit2==1.7.0
Pygments==2.10.0
pygtrie==2.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
python-benedict==0.24.3    
python-dateutil==2.8.2     
python-editor==1.0.4       
python-fsutil==0.5.0       
python-slugify==5.0.2      
pytz==2021.3
pywin32==227
PyYAML==6.0
querystring-parser==1.2.4  
requests==2.26.0
rich==10.12.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6    
scikit-learn==1.0
scipy==1.7.1
shortuuid==1.0.1
shtab==1.4.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
tabulate==0.8.9
text-unidecode==1.3        
threadpoolctl==3.0.0       
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2uritemplate==4.1.1
urllib3==1.26.7
voluptuous==0.12.2
waitress==2.0.0
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.2
xmltodict==0.12.0
yarl==1.7.0
zc.lockfile==2.0
zipp==3.6.0